Added POC for Token based AUthentication
