+++ etc/passwd
@@ -1,0 +1,2 @@
+mkuiper:pbkdf2:sha1:1000$gE6GSSIL$6c8d99caf178c27973655764fd7f00c29cdaa575
+alforbes:pbkdf2:sha1:1000$YCXN5fNr$a2d13e03a3cc47a90c3ae3603b71ffc32f4e35c4
